agents/bom_intelligence/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openpyxl>=3.1.0
networkx>=3.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
streamlit>=1.32.0
plotly>=5.20.0
matplotlib>=3.8.0
